:construction_worker: Se implemento manejo de exepciones y validaciones en todos los metodos que contempla a senador

+++ src/main/java/com/crud/CongresoIUD_DAO/service/impl/SenadorServiceImpl.java
@@ -2,6 +2,7 @@ package com.crud.CongresoIUD_DAO.service.impl;
 
 import com.crud.CongresoIUD_DAO.dto.request.SenadorDTORequest;
 import com.crud.CongresoIUD_DAO.dto.response.SenadorDTO;
+import com.crud.CongresoIUD_DAO.exceptions.UserNotFoundException;
 import com.crud.CongresoIUD_DAO.model.Senador;
 import com.crud.CongresoIUD_DAO.repository.ISenadorRepository;
 import com.crud.CongresoIUD_DAO.service.iface.ISenadorService;
@@ -9,9 +10,8 @@ import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,13 +23,14 @@ public class SenadorServiceImpl implements ISenadorService {
     @Autowired
     public SenadorServiceImpl(ISenadorRepository senadorRepository) {
         this.senadorRepository = senadorRepository;
-
     }
 
-@Override
+    @Override
     public List<SenadorDTO> findAll() {
 
         List<Senador> senadores = senadorRepository.findAll();
+
+        if(senadores.isEmpty()) throw new UserNotFoundException("No hay senadores en la base de datos");
 
         return senadores.stream().map(senador -> SenadorDTO.builder()
                 .id(senador.getId())
@@ -41,40 +42,62 @@ public class SenadorServiceImpl implements ISenadorService {
     }
 
     @Override
-    public SenadorDTO saveSenador(SenadorDTORequest senadorDTORequest) {
+    public String saveSenador(SenadorDTORequest senadorDTORequest)   {
 
         Senador senador;
 
         senador = senadorRepository.findSenadorByNombre(senadorDTORequest.getNombre());
 
-        if(senador == null){
+        Senador existingSenador = senadorRepository.findSenadorByNombre(senadorDTORequest.getNombre());
+        if (existingSenador != null) {
+            throw new IllegalArgumentException("Ya existe el senador " + senadorDTORequest.getNombre() + " en la base de datos");
+        }
+
+        if (senador == null) {
             senador = new Senador();
             senador.setNombre(senadorDTORequest.getNombre());
             senador.setDepartamento(senadorDTORequest.getDepartamento());
             senador.setFecha(senadorDTORequest.getFecha());
             senador.setPartido_politico_id(senadorDTORequest.getPartido_politico_id());
             senadorRepository.save(senador);
-        } else {
+        }
+        return "Se ha guardado el senador " + senadorDTORequest.getNombre() + " con exito";
+    }
+
+    @Override
+    public String update (SenadorDTORequest senadorDTORequest, int senadorId) {
+        Optional<Senador> senadores = this.senadorRepository.findById(Long.valueOf(senadorId));
+
+        if(senadores.isEmpty()){
+            throw new IllegalArgumentException("No existe el senador con id " + senadorId + " en la base de datos");
+        }
+
+            Senador senador = senadores.get();
             senador.setNombre(senadorDTORequest.getNombre());
             senador.setDepartamento(senadorDTORequest.getDepartamento());
             senador.setFecha(senadorDTORequest.getFecha());
             senador.setPartido_politico_id(senadorDTORequest.getPartido_politico_id());
-            senadorRepository.save(senador);
-        }
-        return  SenadorDTO.builder()
-                .id(senador.getId())
-                .nombre(senador.getNombre())
-                .departamento(senador.getDepartamento())
-                .fecha(String.valueOf(senador.getFecha()))
-                .partido_politico_id(String.valueOf(senador.getPartido_politico_id()))
-                .build();
+
+            this.senadorRepository.save(senador);
+
+        return "Se ha actualizado el senador con id " + senadorId + " en la base de datos";
+
+
     }
 
-    @Override
-    public SenadorDTO deleteSenador(long id) {
 
-        senadorRepository.deleteById(id);
-        return SenadorDTO.builder().build();
+    @Override
+    public String deleteSenador(long id) {
+
+        Optional<Senador> senador = senadorRepository.findById(id);
+
+        if (senador.isEmpty()) {
+            throw new UserNotFoundException("No existe el senador con id " + id + " en la base de datos");
+        }else {
+            senadorRepository.deleteById(id);
+        }
+
+        return "Se ha eliminado el senador con id " + id + " con exito";
     }
